requirements.txt
black==25.1.0
blinker==1.9.0
click==8.2.1
Flask==2.2.5
Flask-Babel==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.3
